+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -109,7 +109,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kevxwaj8946tjn61v53bxxqk"
+        "conv_01kexrj7rpgwtpf9mvzswhseaq"
     )
     r"""Unique conversation identifier with `conv_` prefix."""