+++ packages/orq-rc/src/orq_ai_sdk/models/listcontactsop.py
@@ -195,7 +195,7 @@ class ListContactsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-13T14:58:54.165Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-14T08:04:31.906Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")