+++ packages/orq-rc/RELEASES.md
@@ -4649,3 +4649,13 @@ Based on:
 - [python v4.2.0-rc.27] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.27] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.27 - packages/orq-rc
+
+## 2026-01-14 08:07:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.685.0 (2.794.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.28] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.28] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.28 - packages/orq-rc